requirements.txt
cassandra-driver==3.24.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
findspark==1.4.2
geomet==0.2.1.post1
idna==2.10
kafka-python==2.0.1
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
Pillow==7.2.0
psutil==5.7.2
py4j==0.10.9
pyparsing==2.4.7
PySocks==1.7.1
pyspark==3.0.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
spark-nlp==2.6.1
tweepy==3.9.0
urllib3==1.25.10
wordcloud==1.8.0
